requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
python-dotenv==1.0.1
openai==1.12.0
huggingface-hub==0.20.3
transformers==4.37.2
torch==2.2.0
sentencepiece==0.1.99
numpy==1.24.3
pandas==2.2.0
plotly==5.18.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
